+++ init-android-openssl.sh
@@ -16,10 +16,12 @@
 # limitations under the License.
 #
 
-#IJK_OPENSSL_UPSTREAM=https://github.com/openssl/openssl
-IJK_OPENSSL_UPSTREAM=https://github.com/Bilibili/openssl.git
-IJK_OPENSSL_FORK=https://github.com/Bilibili/openssl.git
-IJK_OPENSSL_COMMIT=OpenSSL_1_0_2n
+
+#IJK_OPENSSL_UPSTREAM=https://github.com/Bilibili/openssl.git
+#IJK_OPENSSL_FORK=https://github.com/Bilibili/openssl.git
+IJK_OPENSSL_UPSTREAM=https://github.com/openssl/openssl
+IJK_OPENSSL_FORK=https://github.com/openssl/openssl
+IJK_OPENSSL_COMMIT=OpenSSL_1_1_1-stable
 IJK_OPENSSL_LOCAL_REPO=extra/openssl
 
 set -e
@@ -33,7 +35,7 @@ function pull_fork()
     echo "== pull openssl fork $1 =="
     sh $TOOLS/pull-repo-ref.sh $IJK_OPENSSL_FORK android/contrib/openssl-$1 ${IJK_OPENSSL_LOCAL_REPO}
     cd android/contrib/openssl-$1
-    git checkout ${IJK_OPENSSL_COMMIT} -B ijkplayer
+    git checkout ${IJK_OPENSSL_COMMIT}
     cd -
 }